requirements.txt
django
djangorestframework
geopy
faker